fix: Memisahkan Class Parent dan Child
