+++ basics/15_oop_lanjutan/1_inheritance_Hewan.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Basic\OOP_Lanjutan;
+
 echo "<pre>";
 
 echo "<h1>Contoh Inheritance dalam PHP</h1>";
@@ -51,34 +53,3 @@ echo "Hasil Salah satu method Pada Class Hewan",PHP_EOL;
 $pengguna->salam();
 echo "<br>";
 echo "<br>";
-
-echo "class Kucing extends Hewan 
-{
-    // Method Milik Class Kucing
-    public function bermain() {
-        echo 'Aku Kucing yang suka Bermain';
-    }
-}";
-
-class Kucing extends Hewan 
-{
-    // Method Milik Class Kucing
-    public function bermain() {
-        echo 'Aku Kucing yang suka Bermain';
-    }
-}
-
-echo "<br>";
-echo "<br>";
-echo "Class Kucing merupakan turunan dari Class Hewan",PHP_EOL;
-echo "<br>";
-$kucing = new Kucing;
-echo "Hasil pemanggilan salah satu method class Hewan melalui class Kucing",PHP_EOL;
-$kucing->salam();
-echo "<br>";
-echo "<br>";
-echo "Hasil pemanggilan method bermain pada Class Kucing",PHP_EOL;
-$kucing->bermain();
-  
-echo "</pre>";
-
